Modify IntcodeComputer for Long and complete Day 9 Part I

+++ src/main/kotlin/com/emmav/adventofcode/day13.kt
@@ -7,7 +7,7 @@ private val memory = File("input/day13.txt")
     .readLines()
     .first()
     .split(",")
-    .map { it.toInt() }
+    .map { it.toLong() }
     .toTypedArray()
 
 fun main() {
